Simplify a lot of the math.
For some reason, initial battlefield size was doubled, but the window
size was what you said it should be. So an 800x800 window was reporting
a size of 1600x1600. Your tank's x/y were ALSO up to 1600x1600, so when
you tried to go to the "center" of the screen, you were really going to
the bottom right corner. Wack math is wack. Time to go.